requirements.txt
asgiref==3.2.7
Django==3.0.6
pytz==2020.1
sqlparse==0.3.1
django-recaptcha3==0.2.0
dj-authentication==0.2.0
django-username-email==2.3.0
cuser~=0.0.4
django-phonenumber-field
phonenumberslite
requests~=2.23.0
django_compressor
python-dateutil~=2.8.1
beautifulsoup4~=4.9.1
dateparser~=0.7.6
django-autoslug~=1.9.7
django-allauth
django-crispy-forms